backend/requirements.txt
# Web框架
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# 数据库
mysql-connector-python>=8.0.0
SQLAlchemy>=2.0.0
alembic>=1.12.0
pymysql>=1.1.0

# 工具包
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
PyYAML>=6.0.1
psutil>=5.9.0
ultralytics>=8.0.0
seaborn>=0.12.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0

# 深度学习
torch>=2.0.0
torchvision>=0.15.0

# 数据处理
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.66.0 